Part.1/requirement.txt
flake8==7.3.0
mccabe==0.7.0
numpy==2.2.6
pillow==12.0.0
pycodestyle==2.14.0
pyflakes==3.4.0
